[feat] 로그인 인가

+++ src/main/kotlin/com/example/onlineshoping/project/domain/dto/request/CreateBuyerRequest.kt
@@ -5,5 +5,5 @@ data class CreateBuyerRequest(
     val password:String,
     val name:String,
     val phoneNumber:String,
-    val balance:Int
+    val balance:Long
 )
