+++ src/main/kotlin/com/example/onlineshoping/project/domain/dto/response/BuyerResponse.kt
@@ -6,5 +6,5 @@ data class BuyerResponse(
     val password:String,
     val name:String,
     val phoneNumber:String,
-    val balance:Int
+    val balance:Long
 )
